AbstractLogger renamed to concrete ConsoleFileLogger.
Code cleared.
