+++ Sources/PT.PM.Cli/FileLogger.cs
@@ -1,4 +1,5 @@
-﻿using PT.PM.Common;
+﻿using Newtonsoft.Json;
+using PT.PM.Common;
 using PT.PM.Common.CodeRepository;
 using PT.PM.Common.Exceptions;
 using PT.PM.Matching;
@@ -8,16 +9,18 @@ using System.Threading;
 
 namespace PT.PM.Cli
 {
-    public abstract class AbstractLogger : ILogger
+    public class FileLogger : ILogger
     {
         private int errorCount;
         private string logPath;
 
         public int ErrorCount => errorCount;
 
-        protected virtual NLog.Logger FileLogger => NLog.LogManager.GetLogger("file");
+        protected NLog.Logger FileInternalLogger => NLog.LogManager.GetLogger("file");
 
         protected NLog.Logger ErrorsLogger => NLog.LogManager.GetLogger("errors");
+
+        protected NLog.Logger MatchLogger { get; } = NLog.LogManager.GetLogger("match");
 
         public string LogsDir
         {
@@ -49,7 +52,7 @@ namespace PT.PM.Cli
         public virtual void LogError(string message)
         {
             ErrorsLogger.Error(message);
-            FileLogger.Error(message);
+            FileInternalLogger.Error(message);
             Interlocked.Increment(ref errorCount);
         }
 
@@ -57,16 +60,39 @@ namespace PT.PM.Cli
         {
             var exString = ex.FormatExceptionMessage();
             ErrorsLogger.Error(exString);
-            FileLogger.Error(exString);
+            FileInternalLogger.Error(exString);
             Interlocked.Increment(ref errorCount);
+        }
+
+        public virtual void LogInfo(object infoObj)
+        {
+            string message;
+            var progressEventArgs = infoObj as ProgressEventArgs;
+            if (progressEventArgs != null)
+            {
+                string value = progressEventArgs.Progress >= 1
+                    ? $"{(int)progressEventArgs.Progress} items"
+                    : $"{(int)(progressEventArgs.Progress * 100):0.00}%";
+                message = $"Progress: {value}; File: {progressEventArgs.CurrentFile}";
+                LogInfo(message);
+            }
+            else
+            {
+                var matchingResult = infoObj as MatchingResult;
+                if (matchingResult != null)
+                {
+                    var matchingResultDto = MatchingResultDto.CreateFromMatchingResult(matchingResult, SourceCodeRepository);
+                    var json = JsonConvert.SerializeObject(matchingResultDto, Formatting.Indented);
+                    MatchLogger.Info(json);
+                    LogInfo($"Pattern matched: {Environment.NewLine}{json}{Environment.NewLine}");
+                }
+            }
         }
 
         public virtual void LogInfo(string message)
         {
-            FileLogger.Info(message);
+            FileInternalLogger.Info(message);
         }
-
-        public abstract void LogInfo(object infoObj);
 
         public virtual void LogDebug(string message)
         {
@@ -76,10 +102,8 @@ namespace PT.PM.Cli
 #endif
             if (isDebug || IsLogDebugs)
             {
-                FileLogger.Debug(message);
+                FileInternalLogger.Debug(message);
             }
         }
-
-        protected abstract void LogMatchingResult(MatchingResult matchingResult);
     }
 }
