+++ Sources/PT.PM.Cli/PT.PM.Cli.csproj
@@ -72,8 +72,8 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AbstractLogger.cs" />
-    <Compile Include="ConsoleLogger.cs" />
+    <Compile Include="FileLogger.cs" />
+    <Compile Include="ConsoleFileLogger.cs" />
     <Compile Include="WorkflowLoggerHelper.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
